HBASE-22548 Split TestAdmin1
